pyproject.toml
[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "federated-health-risk"
version = "0.1.0"
description = "End-to-end MLOps system for federated health risk prediction"
authors = ["Sahal Saeed <sahal@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2"
polars = "^1.6"
numpy = "^1.26"
scikit-learn = "^1.5"
torch = {version = "^2.4", source = "pytorch"}
torchvision = {version = "^0.19", source = "pytorch"}
torchaudio = {version = "^2.4", source = "pytorch"}
flwr = "^1.11"
mlflow = "^2.15"
prefect = "^3.0"
fastapi = "^0.115"
uvicorn = "^0.32"
streamlit = "^1.39"
great-expectations = "^1.1"
sentence-transformers = "^3.1"
alibi-detect = "^0.12"
river = "^0.21"
pydantic = "^2.9"
folium = "^0.16"
plotly = "^5.24"
geopandas = "^1.0"
shapely = "^2.0"
kafka-python = "^2.0"
requests = "^2.32"
typer = "^0.12"
hydra-core = "^1.3"
python-dotenv = "^1.0"
sqlalchemy = "^2.0"
pyarrow = "^16.1"
kagglehub = "^0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^5.0"
ruff = "^0.6"
black = "^24.8"
pre-commit = "^3.8"

[tool.pytest.ini_options]
pythonpath = ["src"]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"
